Se añadio el user storie de añadir artista y buscar artistas totales y por nombre

+++ src/main/java/com/backend/dtos/creates/CreateArtistaDto.java
@@ -10,8 +10,8 @@ public class CreateArtistaDto {
     private String foto_perfil;
     private String foto_portada;
     private String biografia;
-    private Integer seguidores;
-    private Integer oyentesTotal;
+    private Long seguidores;
+    private Long oyentesTotal;
 	private String departamento_origen;
 	private String genero_musical;
 }
